Rename file according to asyncComponent function name
