+++ ui/apps/platform/src/Containers/MainPage/Body.tsx
@@ -59,7 +59,7 @@ import { IsFeatureFlagEnabled } from 'hooks/useFeatureFlags';
 import useAnalytics from 'hooks/useAnalytics';
 import { selectors } from 'reducers';
 
-import asyncComponent from './AsyncComponent';
+import asyncComponent from './asyncComponent';
 import InviteUsersModal from './InviteUsers/InviteUsersModal';
 
 function NotFoundPage(): ReactElement {
