+++ ui/apps/platform/src/Containers/MainPage/asyncComponent.tsx
